[SPARK-131458] -Pull credential dump up to SparkHadoopUtil; It'd broader than just YARN. -update docs to clarify HBase/Hive tokens are entirely optional. -general review of text.
