+++ yarn/src/main/scala/org/apache/spark/deploy/yarn/YarnSparkHadoopUtil.scala
@@ -17,16 +17,13 @@
 
 package org.apache.spark.deploy.yarn
 
-import java.io.{File, IOException}
+import java.io.File
 import java.lang.reflect.UndeclaredThrowableException
 import java.nio.charset.StandardCharsets.UTF_8
 import java.security.PrivilegedExceptionAction
-import java.text.DateFormat
-import java.util.Date
 import java.util.regex.Matcher
 import java.util.regex.Pattern
 
-import scala.collection.JavaConverters._
 import scala.collection.mutable.HashMap
 import scala.reflect.runtime._
 import scala.util.Try
@@ -39,7 +36,6 @@ import org.apache.hadoop.mapred.{JobConf, Master}
 import org.apache.hadoop.security.Credentials
 import org.apache.hadoop.security.UserGroupInformation
 import org.apache.hadoop.security.token.{Token, TokenIdentifier}
-import org.apache.hadoop.security.token.delegation.AbstractDelegationTokenIdentifier
 import org.apache.hadoop.yarn.api.ApplicationConstants
 import org.apache.hadoop.yarn.api.ApplicationConstants.Environment
 import org.apache.hadoop.yarn.api.records.{ApplicationAccessType, ContainerId, Priority}
@@ -330,48 +326,6 @@ class YarnSparkHadoopUtil extends SparkHadoopUtil {
     }
   }
 
-  /**
-   * Dump the credentials' tokens to string values.
-   * @param credentials credentials
-   * @return an iterator over the string values. If no credentials are passed in: an empty list
-   */
-  def dumpTokens(credentials: Credentials): Iterable[String] = {
-    if (credentials != null) {
-      credentials.getAllTokens.asScala.map(tokenToString)
-    } else {
-      Seq()
-    }
-  }
-
-  /**
-   * Convert a token to a string.
-   * If its an abstract delegation token, attempt to unmarshall it and then
-   * print more details, including timestamps in human-readable form.
-   * @param token token to convert to a string
-   * @return a printable string value.
-   */
-  def tokenToString(token: Token[_ <: TokenIdentifier]): String = {
-    val df = DateFormat.getDateTimeInstance(DateFormat.SHORT, DateFormat.SHORT)
-    val buffer = new StringBuilder(128)
-    buffer.append(token.toString)
-    try {
-      val ti = token.decodeIdentifier
-      buffer.append("; ").append(ti)
-      ti match {
-        case dt: AbstractDelegationTokenIdentifier =>
-          // include human times and the renewer, which the HDFS tokens toString omits
-          buffer.append(s"; Renewer: ${dt.getRenewer}")
-          buffer.append("; Issued: ").append(df.format(new Date(dt.getIssueDate)))
-          buffer.append("; Max Date: ").append(df.format(new Date(dt.getMaxDate)))
-        case _ =>
-      }
-    } catch {
-      case e: IOException => {
-        logDebug("Failed to decode $token: $e", e)
-      }
-    }
-    buffer.toString
-  }
 }
 
 object YarnSparkHadoopUtil {
